feat(playerconnect): add link validation and rendering in chat
- Add isValid method to PlayerConnectLink for URL validation
- Implement PlayerConnectRenderer for reusable room display components
- Modify ChatOverlay to render player connect links in messages
- Refactor PlayerConnectJoinInjector to use new renderer
